Use safer topics.
This project got a little extra ambitious since I was converting from
Text to Filter in the parser anyway and we had two different types of
things using the same data.

The older code would just make a bunch of uniform things that would be
used vastly differently where different fields had different
meanings.  Now we parse the config into things we actually want and
have separate types for the separate runtimes.

The one caveat is that we don't parse InfluxDB queries directly into
the Query type.  The primary reason for this is because they don't
have an Eq instance and this make testing annoying.  There are a few
ways around that, but it's not a lot of value a this point since the
query parsing code doesn't seem to have any validation anyway.